+++ designer/client/cypress/e2e/fragment.cy.ts
@@ -38,6 +38,7 @@ describe("Fragment", () => {
         cy.get("[id$='option-1']").click({ force: true });
 
         // Display Add list item errors when blank value
+        cy.get("[data-testid='settings:4']").contains("User defined list").click();
         cy.get("[data-testid='settings:4']").find("[id='ace-editor']").type("{enter}");
         cy.get("[data-testid='settings:4']").find("[data-testid='form-helper-text']").should("be.visible");
         cy.get("[data-testid='settings:4']").contains("This field is mandatory and can not be empty");
@@ -78,6 +79,7 @@ describe("Fragment", () => {
         toggleSettings(5);
         cy.get("[data-testid='settings:5']").contains("Any value").click();
         cy.get("[id$='option-0']").click({ force: true });
+        cy.get("[data-testid='settings:5']").contains("User defined list").click();
         cy.get("[data-testid='settings:5']").find("[id='ace-editor']").type("#meta.processName");
         cy.get("[data-testid='settings:5']").contains("Typing...").should("not.exist");
         cy.get("[data-testid='settings:5']").find("[id='ace-editor']").type("{enter}");
@@ -119,6 +121,44 @@ describe("Fragment", () => {
 
         cy.get("@window").find("[data-testid='settings:6']").matchImage();
 
+        // Provide String Fixed value inputMode
+        cy.get("@window").contains("+").click();
+        cy.get("[data-testid='fieldsRow:7']").find("[placeholder='Field name']").type("any_value_with_suggestions_preset");
+        toggleSettings(7);
+
+        // Select any value with suggestions Input mode
+        cy.get("[data-testid='settings:7']").contains("Any value").click();
+        cy.get("[id$='option-1']").click({ force: true });
+
+        // Activate preset mode
+        cy.get("[data-testid='settings:7']").contains("Preset").click();
+
+        // Initial value should be disabled when there is no preset selected
+        cy.get("[data-testid='settings:7']")
+            .find("label")
+            .contains(/initial value/i)
+            .siblings()
+            .find("input")
+            .should("be.disabled");
+
+        // Select available values from Preset
+        cy.get("[data-testid='settings:7']")
+            .find("label")
+            .contains(/preset selection/i)
+            .siblings()
+            .eq(0)
+            .click();
+        cy.get("[id$='option-1']").click({ force: true });
+
+        // Select Initial value
+        cy.get("[data-testid='settings:7']")
+            .find("label")
+            .contains(/initial value/i)
+            .siblings()
+            .eq(0)
+            .click();
+        cy.get("[id$='option-1']").click({ force: true });
+
         cy.get("@window")
             .contains(/^apply$/i)
             .click();
@@ -143,7 +183,7 @@ describe("Fragment", () => {
         cy.intercept("POST", "/api/nodes/*/validation").as("fragmentInputValidation");
 
         cy.get("[model-id^=e2e][model-id$=fragment-test-process]").should("be.visible").trigger("dblclick");
-        cy.get("#nk-graph-fragment [model-id='input']").should("be.visible");
+        cy.get("#nk-graph-fragment [model-id='input']").scrollIntoView().should("be.visible");
 
         cy.wait("@fragmentInputValidation");
         cy.get("[data-testid=window]").matchImage();
@@ -152,6 +192,24 @@ describe("Fragment", () => {
         cy.get('[title="name_string_fixed"]').siblings().eq(0).contains("#meta.processName");
         cy.get('[title="name_string_fixed"]').find('[title="Hint text test"]').should("be.visible");
         cy.get('[title="non_boolean_or_string"]').siblings().eq(0).contains("1");
+
+        // any value with suggestions preset verification
+        cy.get('[title="any_value_with_suggestions_preset"]').siblings().eq(0).as("anyValueWithSuggestionField");
+
+        cy.get("@anyValueWithSuggestionField").find("input").should("have.value", "Email Marketing 12.2019");
+        cy.get("@anyValueWithSuggestionField").clear().type("Campaign 2020");
+        cy.get("[id$='option-0']").click({ force: true });
+        cy.get("@anyValueWithSuggestionField").find("input").should("have.value", "Campaign 2020 News");
+        cy.get("@anyValueWithSuggestionField").find('[title="Switch to expression mode"]').click();
+        cy.get("@anyValueWithSuggestionField").contains('{"key":"9d6d4e3e-0ba6-43bb-8696-58432e8f6bd8","label":"Campaign ');
+        cy.get("@anyValueWithSuggestionField").find("#ace-editor").type("{selectall}t");
+        cy.get("@anyValueWithSuggestionField")
+            .find('[title="Expression must be valid JSON to switch to basic mode"]')
+            .should("be.disabled");
+        cy.get("@anyValueWithSuggestionField")
+            .find("#ace-editor")
+            .type("{selectall}{backspace}")
+            .type('{"key": "9d6d4e3e-0ba6-43bb-8696-58432e8f6bd8", "label": "Campaign 2020 News"}', { parseSpecialCharSequences: false });
 
         cy.get("[data-testid=window]").find("input[value=testOutput]").type("{selectall}fragmentResult");
         cy.contains(/^apply/i)
@@ -198,6 +256,13 @@ describe("Fragment", () => {
         cy.contains(/^save\*$/i).click();
         cy.contains(/^ok$/i).click();
 
+        // Verify if Frontend received correct data after save
+        cy.get("[model-id^=e2e][model-id$=fragment-test-process]").trigger("dblclick");
+        cy.get('[title="any_value_with_suggestions_preset"]').siblings().eq(0).find("input").should("have.value", "Campaign 2020 News");
+        cy.contains(/^apply/i)
+            .should("be.enabled")
+            .click();
+
         // Go back to the fragment
         cy.go(-1);
 
@@ -205,7 +270,7 @@ describe("Fragment", () => {
         cy.get("[model-id=input]").should("be.visible").trigger("dblclick");
         cy.get("[data-testid=window]").should("be.visible").as("window");
         cy.get("@window").contains("+").click();
-        cy.get("[data-testid='fieldsRow:7']").find("[placeholder='Field name']").type("test5");
+        cy.get("[data-testid='fieldsRow:8']").find("[placeholder='Field name']").type("test5");
         cy.get("@window")
             .contains(/^apply$/i)
             .click();