+++ designer/client/src/components/graph/node-modal/fragment-input-definition/settings/variants/StringBooleanVariants/FixedListVariant.tsx
@@ -1,28 +1,26 @@
 import React from "react";
 import { useTranslation } from "react-i18next";
-import { FixedListParameterVariant, FixedValuesType, onChangeType } from "../../../item";
+import { FixedListParameterVariant, onChangeType } from "../../../item";
 import InitialValue from "../fields/InitialValue";
 import { FixedValuesSetting } from "../fields/FixedValuesSetting";
 import { SettingLabelStyled } from "../fields/StyledSettingsComponnets";
 import { TextAreaNodeWithFocus } from "../../../../../../withFocus";
-import { FixedValuesPresets, NodeValidationError, VariableTypes } from "../../../../../../../types";
+import { NodeValidationError, VariableTypes } from "../../../../../../../types";
 import { getValidationErrorsForField } from "../../../../editors/Validators";
 import { FormControl } from "@mui/material";
+import { FixedValuesGroup } from "../fields/FixedValuesGroup";
 
 interface Props {
     item: FixedListParameterVariant;
     onChange: (path: string, value: onChangeType) => void;
     path: string;
-    fixedValuesPresets: FixedValuesPresets;
     readOnly: boolean;
     variableTypes: VariableTypes;
     errors: NodeValidationError[];
 }
 
-export const FixedListVariant = ({ item, path, onChange, fixedValuesPresets, readOnly, variableTypes, errors }: Props) => {
+export const FixedListVariant = ({ item, path, onChange, readOnly, variableTypes, errors }: Props) => {
     const { t } = useTranslation();
-
-    const presetListItemOptions = fixedValuesPresets?.[item.fixedValuesListPresetId] ?? [];
 
     const fixedValuesList = item.valueEditor.fixedValuesList ?? [];
 
@@ -30,15 +28,14 @@ export const FixedListVariant = ({ item, path, onChange, fixedValuesPresets, rea
 
     return (
         <>
-            {/*<FixedValuesGroup fixedValuesType={fixedValuesType} path={path} onChange={onChange} readOnly={readOnly} />*/}
+            <FixedValuesGroup fixedValuesType={fixedValuesType} path={path} onChange={onChange} readOnly={readOnly} item={item} />
             <FixedValuesSetting
                 path={path}
                 onChange={onChange}
                 fixedValuesType={fixedValuesType}
                 presetSelection={item.presetSelection}
                 fixedValuesList={fixedValuesList}
-                fixedValuesPresets={fixedValuesPresets}
-                fixedValuesListPresetId={item.fixedValuesListPresetId}
+                dictId={item.valueEditor.dictId}
                 readOnly={readOnly}
                 variableTypes={variableTypes}
                 errors={errors}
@@ -50,7 +47,7 @@ export const FixedListVariant = ({ item, path, onChange, fixedValuesPresets, rea
                 path={path}
                 item={item}
                 onChange={onChange}
-                options={fixedValuesType === FixedValuesType.ValueInputWithFixedValuesProvided ? fixedValuesList : presetListItemOptions}
+                options={fixedValuesList}
                 readOnly={readOnly}
                 variableTypes={variableTypes}
                 fieldErrors={getValidationErrorsForField(errors, `$param.${item.name}.$initialValue`)}